+++ build_runner/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 2.8.1-wip
 
+- Improved warnings when an option is specified for an unknown builder.
 - Bug fix: require `args` 2.5.0.
 
 ## 2.8.0